UPDATE: getSize int -> long

+++ src/main/java/mops/persistence/FileInfo.java
@@ -28,5 +28,5 @@ public interface FileInfo {
    /**
     * @return byte size.
     */
-   int getSize();
+   long getSize();
 }
